requirements.txt
Django==2.1
psycopg2
djangorestframework
python-dotenv
djangorestframework_jwt
django-cors-headers
django-filter
statsmodels
scipy
pygments